% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cortest.R
\name{cortest}
\alias{cortest}
\title{Cortest}
\usage{
cortest(
  datasource,
  var1 = "CO2Flux",
  var2,
  interval = TRUE,
  cormethod = "pearson",
  na.rm = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{datasource}{a dataframe or tibble with at least two columns to be able to correlate.}

\item{var1}{character vector for, generally, a constant variable to be correlated with  dependent variables during function call}

\item{var2}{character vector for the variable of interest}

\item{interval}{logical value. By default, \strong{TRUE}. Displays confidence interval on \code{ggscatter()} plot}

\item{cormethod}{character vector specifying correlation method. By default, \strong{pearson}. Allows for use of 1 of 3 correlation methods: \strong{pearson, kendall, or spearman}. Tests for association between paired samples.}

\item{na.rm}{a logical indicating whether missing values should be removed}

\item{verbose}{a logical indicating whether to use verbose for flow of information}
}
\value{
outputs an \code{htest} object, which is a list containing information about null and alternative hypotheses, estimated distribution parameters, the test-statistic, the p-value, and optional confidence intervals for distribution parameters.
}
\description{
Cortest
}
\examples{
cortest(datasource = Date, var2 = "Total_N", cormethod = "pearson", verbose = TRUE)
}
